requirements.txt
asgiref==3.6.0
dj-database-url==2.0.0
Django==4.2
django-cors-headers==3.14.0
django-filter==23.1
djangorestframework==3.14.0
gunicorn==20.1.0
importlib-metadata==6.4.1
Markdown==3.4.3
psycopg2-binary==2.9.6
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.3
typing_extensions==4.5.0
tzdata==2023.3
zipp==3.15.0
